+++ src/components/Post/PostDetail/Post.tsx
