+++ src/components/Post/PostDetail/PostTemplate.tsx
