+++ src/components/Post/PostDetail/index.ts
@@ -1,0 +1,5 @@
+import Post from './Post';
+export { default as PostSkeleton } from './PostSkeleton';
+export { default as PostTemplate } from './PostTemplate';
+
+export default Post;
