+++ src/components/Post/PostList/index.ts
@@ -1,0 +1,6 @@
+import PostList from './PostList';
+export { default as PostListSkeleton } from './ListSkeleton';
+export { default as TagList } from './TagList';
+export { default as SearchBar } from './SearchBar';
+
+export default PostList;
